\name{BayesMAMS-package}
\alias{BayesMAMS-package}
\alias{BayesMAMS}
\docType{package}
\title{
Designing Bayesian Multi-Arm Multi-Stage Studies
}
\description{
Calculating Bayesian sample sizes for multi-arm trials where several experimental treatments are compared to a common control, perhaps even at multiple stages.
}
\details{
\tabular{ll}{
Package: \tab BayesMAMS\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-11-02\cr
License: \tab GPL (>= 2)\cr
}
Function \code{ssbayes} finds Bayesian sample sizes for multi-arm trials given some criteria described in Whitehead et al. (2015).
}
\author{
Philip Pallmann, Amanda Turner

Maintainer: Philip Pallmann <p.pallmann@lancaster.ac.uk>
}
\references{
Whitehead, J., Cleary, F., Turner, A. (2015) Bayesian sample sizes for exploratory clinical trials comparing multiple experimental treatments with a control. Statistics in Medicine, 34(12), 2048--2061.
}
\keyword{package}
